refactor: move source into src and use path alias
