+++ src/components/PhraseRotator.tsx
