+++ src/components/SettingsModal.tsx
@@ -2,13 +2,13 @@ import React, { useState, useEffect, useRef } from 'react';
 import { motion, Variants } from 'framer-motion';
 import Modal from './Modal';
 import Button from './Button';
-import { NotificationType } from '../types';
-import * as Storage from '../services/localStorageService';
+import { NotificationType } from '@/types';
+import * as Storage from '@/services/localStorageService';
 import { 
     ArrowDownTrayIcon, ArrowUpTrayIcon, SparklesIcon, TrashIcon, CogIcon,
     AdjustmentsHorizontalIcon, SquaresPlusIcon, PaletteIcon
 } from './icons';
-import { ThemeName } from '../styles/theme';
+import { ThemeName } from '@/styles/theme';
 
 const sectionContainerVariants: Variants = {
   hidden: { opacity: 1 }, // Parent is visible
