+++ src/motion/variants.ts
