+++ src/services/localStorageService.test.ts
@@ -3,7 +3,7 @@
 import { GenerateContentResponse as GenAIResponse, GenerateContentParameters as GenAIParameters } from '@google/genai';
 import { describe, test, expect, beforeEach, jest } from '@jest/globals';
 import * as localStorageServiceOriginal from './localStorageService'; // Adjust path
-import { Project, Idea } from '../types'; // Adjust path
+import { Project, Idea } from '@/types'; // Adjust path
 // Import types from @google/genai directly for mocking clarity
 
 // Declare the mock function with the correct signature
