+++ vite.config.ts
@@ -1,6 +1,7 @@
 /// <reference types="node" />
 
 import { defineConfig, loadEnv } from 'vite';
+import path from 'path';
 import react from '@vitejs/plugin-react';
 
 // https://vitejs.dev/config/
@@ -10,6 +11,11 @@ export default defineConfig(({ mode }) => {
   const env = loadEnv(mode, '.', '');
   return {
     plugins: [react()],
+    resolve: {
+      alias: {
+        '@': path.resolve(__dirname, 'src')
+      }
+    },
     define: {
       // Expose environment variables to the client.
       // This makes `process.env.API_KEY` available in the code.
